backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
google-generativeai>=0.8.0
passlib[bcrypt]
pyjwt
python-multipart
pydantic[email]
